refactoring code in to modules
