+++ index.html
@@ -4,7 +4,7 @@
   <meta charset="UTF-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="./style.css">
+  <link rel="stylesheet" href="./index.css">
   <title>Awesome Books Store</title>
 </head>
 <body>
@@ -46,6 +46,6 @@
   <footer class="footer">
     <p class="copirgh"> &copy; 2024 Awsome Books Addis ababa, Inc. All rights reserverd</p>
   </footer>
-  <script src="./index.js"></script>
+  <script type="module" src="./index.js"></script>
 </body>
 </html>
