+++ modules/Store.js
@@ -1,4 +1,4 @@
-class Store {
+export default class Store {
   static getBooks() {
     let books;
     if(localStorage.getItem('books') === null) {
@@ -27,5 +27,3 @@ class Store {
     localStorage.setItem('books', JSON.stringify(books));
   }
 }
-
-export {Store};
