+++ package.json
@@ -1,34 +1,25 @@
 {
-  "name": "my-portfolio",
+  "type": "module",
+  "name": "awesome-books",
   "version": "1.0.0",
-  "description": "This is My portfolio page",
+  "description": "<a name=\"readme-top\"></a> <!-- TABLE OF CONTENTS -->",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/Paul-tes/My-Portfolio.git"
+    "url": "git+https://github.com/Paul-tes/Awesome-books.git"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "bugs": {
-    "url": "https://github.com/Paul-tes/My-Portfolio/issues"
+    "url": "https://github.com/Paul-tes/Awesome-books/issues"
   },
-  "homepage": "https://github.com/Paul-tes/My-Portfolio#readme",
-  "devDependencies": {
-    "babel-eslint": "^10.1.0",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-plugin-import": "^2.27.5",
-    "hint": "^2.0.0",
-    "stylelint": "^13.13.1",
-    "stylelint-config-standard": "^21.0.0",
-    "stylelint-csstree-validator": "^1.9.0",
-    "stylelint-scss": "^3.21.0"
-  },
+  "homepage": "https://github.com/Paul-tes/Awesome-books#readme",
   "dependencies": {
-    "jsdom": "^21.1.1"
-  }
+    "luxon": "^3.3.0"
+  },
+  "devDependencies": {}
 }
